Able to train and inference

+++ bevdepth/layers/backbones/bevstereo_lss_fpn.py
@@ -786,9 +786,9 @@ class BEVStereoLSSFPN(BaseLSSFPN):
         img_feat_with_depth = img_feat_with_depth.permute(0, 1, 3, 4, 5, 2)
         geom_xyz = ((geom_xyz - (self.voxel_coord - self.voxel_size / 2.0)) /
                     self.voxel_size).int()
-        feature_map = voxel_pooling_train(
-            geom_xyz,
-            img_feat_with_depth.contiguous().float(), self.voxel_num.cuda())
+        feature_map = voxel_pooling_train(geom_xyz,
+                                          img_feat_with_depth.contiguous(),
+                                          self.voxel_num.cuda())
         if is_return_depth:
             return feature_map.contiguous(), depth
         return feature_map.contiguous()
